refactor: split refactor.py into subpackage with focused modules
