+++ src/prothon/refactor/__init__.py
@@ -1,0 +1,33 @@
+from __future__ import annotations
+
+from prothon.refactor.discovery import discover_drift
+from prothon.refactor.metrics import (
+    collect_cross_module_similarities,
+    collect_module_metrics,
+    collect_pattern_usage,
+)
+from prothon.refactor.models import (
+    DriftCategory,
+    DriftFinding,
+    ModuleMetrics,
+    PatternOccurrence,
+    PatternType,
+    Severity,
+    SimilarityGroup,
+)
+from prothon.refactor.promise_gen import generate_refactor_promise
+
+__all__ = [
+    "DriftCategory",
+    "DriftFinding",
+    "ModuleMetrics",
+    "PatternOccurrence",
+    "PatternType",
+    "Severity",
+    "SimilarityGroup",
+    "collect_cross_module_similarities",
+    "collect_module_metrics",
+    "collect_pattern_usage",
+    "discover_drift",
+    "generate_refactor_promise",
+]
